dgpu: add dgpu.Interface which is Dusk's custom graphics API
This effectively copies the WebGPU/mach-gpu `gpu.Interface` and instead
lets Dusk define it. For now, it's almost a 1:1 copy and almost entirely
commented out.

The reason to do this the same way as `gpu.Interface` is that we should
keep in mind C ABI compatibility, and the ability to intercept GPU API
calls for the purposes of debugging tooling in the future.

We should carefully review which portions of the WebGPU API we want to
keep, and which ones we want to change for a render graph API.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>