msl: fix matching bindings with default pipeline layout
msl.gen() would be invoked with a `bindings` set that matches the desired `entrypoint`,
e.g. it would be invoked twice, once for the vertex shader with a single binding, and once
for the fragment shader with two bindings. That seems correct, but we would discover both
entrypoints and so e.g. in the vertex shader generation call we would crawl `frag_main`
and find the fragment bindings - but those were not passed into the generate function -
leading to a failure to find the bindings.

This fixes the issue by only crawling the desired entrypoint for bindings.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>